refactor view purchased course page into courses folder
