+++ src/router.js
@@ -40,7 +40,7 @@ const routes = [
   },
   {
     path: '/courses/:id',
-    Component: lazy(() => import('pages/courses/view')),
+    Component: lazy(() => import('pages/courses/browse')),
     exact: true,
   },
   {
@@ -187,12 +187,12 @@ const routes = [
   },
   {
     path: '/admin/course-content-management/:id',
-    Component: lazy(() => import('pages/student/dashboard/courses/view')),
+    Component: lazy(() => import('pages/courses/view')),
     exact: true,
   },
   {
     path: '/admin/course-content-management/:courseId/view-lesson/:lessonId',
-    Component: lazy(() => import('pages/student/dashboard/courses/view-lesson')),
+    Component: lazy(() => import('pages/courses/view-lesson')),
     exact: true,
   },
   {
@@ -273,12 +273,12 @@ const routes = [
   },
   {
     path: '/sensei/courses/view/:id',
-    Component: lazy(() => import('pages/student/dashboard/courses/view')),
+    Component: lazy(() => import('pages/courses/view')),
     exact: true,
   },
   {
     path: '/sensei/courses/:courseId/view-lesson/:lessonId',
-    Component: lazy(() => import('pages/student/dashboard/courses/view-lesson')),
+    Component: lazy(() => import('pages/courses/view-lesson')),
     exact: true,
   },
   {
@@ -364,12 +364,12 @@ const routes = [
   },
   {
     path: '/student/dashboard/courses/:id',
-    Component: lazy(() => import('pages/student/dashboard/courses/view')),
+    Component: lazy(() => import('pages/courses/view')),
     exact: true,
   },
   {
     path: '/student/dashboard/courses/:courseId/view-lesson/:lessonId',
-    Component: lazy(() => import('pages/student/dashboard/courses/view-lesson')),
+    Component: lazy(() => import('pages/courses/view-lesson')),
     exact: true,
   },
   {
